+++ node/src/main/kotlin/net/corda/node/services/network/PersistentNetworkMapCache.kt
@@ -31,7 +31,6 @@ import net.corda.node.services.messaging.sendRequest
 import net.corda.node.services.network.NetworkMapService.FetchMapResponse
 import net.corda.node.services.network.NetworkMapService.SubscribeResponse
 import net.corda.node.utilities.AddOrRemove
-import net.corda.node.utilities.DatabaseTransactionManager
 import net.corda.node.utilities.bufferUntilDatabaseCommit
 import net.corda.node.utilities.wrapWithDatabaseTransaction
 import org.hibernate.Session
@@ -39,7 +38,6 @@ import rx.Observable
 import rx.subjects.PublishSubject
 import java.security.PublicKey
 import java.security.SignatureException
-import java.time.Duration
 import java.util.*
 import javax.annotation.concurrent.ThreadSafe
 import kotlin.collections.HashMap
@@ -93,7 +91,7 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
 
     init {
         loadFromFiles()
-        serviceHub.database.transaction { loadFromDB() }
+        serviceHub.database.transaction { loadFromDB(session) }
     }
 
     private fun loadFromFiles() {
@@ -102,7 +100,7 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
     }
 
     override fun getPartyInfo(party: Party): PartyInfo? {
-        val nodes = serviceHub.database.transaction { queryByIdentityKey(party.owningKey) }
+        val nodes = serviceHub.database.transaction { queryByIdentityKey(session, party.owningKey) }
         if (nodes.size == 1 && nodes[0].isLegalIdentity(party)) {
             return PartyInfo.SingleNode(party, nodes[0].addresses)
         }
@@ -117,9 +115,9 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
     }
 
     override fun getNodeByLegalName(name: CordaX500Name): NodeInfo? = getNodesByLegalName(name).firstOrNull()
-    override fun getNodesByLegalName(name: CordaX500Name): List<NodeInfo> = serviceHub.database.transaction { queryByLegalName(name) }
+    override fun getNodesByLegalName(name: CordaX500Name): List<NodeInfo> = serviceHub.database.transaction { queryByLegalName(session, name) }
     override fun getNodesByLegalIdentityKey(identityKey: PublicKey): List<NodeInfo> =
-            serviceHub.database.transaction { queryByIdentityKey(identityKey) }
+            serviceHub.database.transaction { queryByIdentityKey(session, identityKey) }
 
     override fun getNodeByLegalIdentity(party: AbstractParty): NodeInfo? {
         val wellKnownParty = serviceHub.identityService.wellKnownPartyFromAnonymous(party)
@@ -128,9 +126,9 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
         }
     }
 
-    override fun getNodeByAddress(address: NetworkHostAndPort): NodeInfo? = serviceHub.database.transaction { queryByAddress(address) }
+    override fun getNodeByAddress(address: NetworkHostAndPort): NodeInfo? = serviceHub.database.transaction { queryByAddress(session, address) }
 
-    override fun getPeerCertificateByLegalName(name: CordaX500Name): PartyAndCertificate? = serviceHub.database.transaction { queryIdentityByLegalName(name) }
+    override fun getPeerCertificateByLegalName(name: CordaX500Name): PartyAndCertificate? = serviceHub.database.transaction { queryIdentityByLegalName(session, name) }
 
     override fun track(): DataFeed<List<NodeInfo>, MapChange> {
         synchronized(_changed) {
@@ -204,7 +202,7 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
         synchronized(_changed) {
             registeredNodes.remove(node.legalIdentities.first().owningKey)
             serviceHub.database.transaction {
-                removeInfoDB(node)
+                removeInfoDB(session, node)
                 changePublisher.onNext(MapChange.Removed(node))
             }
         }
@@ -238,10 +236,8 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
     }
 
     override val allNodes: List<NodeInfo>
-        get () = serviceHub.database.transaction {
-            createSession {
-                getAllInfos(it).map { it.toNodeInfo() }
-            }
+        get() = serviceHub.database.transaction {
+            getAllInfos(session).map { it.toNodeInfo() }
         }
 
     private fun processRegistration(reg: NodeRegistration) {
@@ -259,10 +255,6 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
     // Changes related to NetworkMap redesign
     // TODO It will be properly merged into network map cache after services removal.
 
-    private inline fun <T> createSession(block: (Session) -> T): T {
-        return DatabaseTransactionManager.current().session.let { block(it) }
-    }
-
     private fun getAllInfos(session: Session): List<NodeInfoSchemaV1.PersistentNodeInfo> {
         val criteria = session.criteriaBuilder.createQuery(NodeInfoSchemaV1.PersistentNodeInfo::class.java)
         criteria.select(criteria.from(NodeInfoSchemaV1.PersistentNodeInfo::class.java))
@@ -272,23 +264,21 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
     /**
      * Load NetworkMap data from the database if present. Node can start without having NetworkMapService configured.
      */
-    private fun loadFromDB() {
+    private fun loadFromDB(session: Session) {
         logger.info("Loading network map from database...")
-        createSession {
-            val result = getAllInfos(it)
-            for (nodeInfo in result) {
-                try {
-                    logger.info("Loaded node info: $nodeInfo")
-                    val node = nodeInfo.toNodeInfo()
-                    addNode(node)
-                    _loadDBSuccess = true // This is used in AbstractNode to indicate that node is ready.
-                } catch (e: Exception) {
-                    logger.warn("Exception parsing network map from the database.", e)
-                }
+        val result = getAllInfos(session)
+        for (nodeInfo in result) {
+            try {
+                logger.info("Loaded node info: $nodeInfo")
+                val node = nodeInfo.toNodeInfo()
+                addNode(node)
+                _loadDBSuccess = true // This is used in AbstractNode to indicate that node is ready.
+            } catch (e: Exception) {
+                logger.warn("Exception parsing network map from the database.", e)
             }
-            if (loadDBSuccess) {
-                _registrationFuture.set(null) // Useful only if we don't have NetworkMapService configured so StateMachineManager can start.
-            }
+        }
+        if (loadDBSuccess) {
+            _registrationFuture.set(null) // Useful only if we don't have NetworkMapService configured so StateMachineManager can start.
         }
     }
 
@@ -313,11 +303,9 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
         }
     }
 
-    private fun removeInfoDB(nodeInfo: NodeInfo) {
-        createSession {
-            val info = findByIdentityKey(it, nodeInfo.legalIdentitiesAndCerts.first().owningKey).single()
-            it.remove(info)
-        }
+    private fun removeInfoDB(session: Session, nodeInfo: NodeInfo) {
+        val info = findByIdentityKey(session, nodeInfo.legalIdentitiesAndCerts.first().owningKey).single()
+        session.remove(info)
     }
 
     private fun findByIdentityKey(session: Session, identityKey: PublicKey): List<NodeInfoSchemaV1.PersistentNodeInfo> {
@@ -328,48 +316,40 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
         return query.resultList
     }
 
-    private fun queryByIdentityKey(identityKey: PublicKey): List<NodeInfo> {
-        createSession {
-            val result = findByIdentityKey(it, identityKey)
-            return result.map { it.toNodeInfo() }
-        }
+    private fun queryByIdentityKey(session: Session, identityKey: PublicKey): List<NodeInfo> {
+        val result = findByIdentityKey(session, identityKey)
+        return result.map { it.toNodeInfo() }
     }
 
-    private fun queryIdentityByLegalName(name: CordaX500Name): PartyAndCertificate? {
-        createSession {
-            val query = it.createQuery(
-                    // We do the JOIN here to restrict results to those present in the network map
-                    "SELECT DISTINCT l FROM ${NodeInfoSchemaV1.PersistentNodeInfo::class.java.name} n JOIN n.legalIdentitiesAndCerts l WHERE l.name = :name",
-                    NodeInfoSchemaV1.DBPartyAndCertificate::class.java)
-            query.setParameter("name", name.toString())
-            val candidates = query.resultList.map { it.toLegalIdentityAndCert() }
-            // The map is restricted to holding a single identity for any X.500 name, so firstOrNull() is correct here.
-            return candidates.firstOrNull()
-        }
+    private fun queryIdentityByLegalName(session: Session, name: CordaX500Name): PartyAndCertificate? {
+        val query = session.createQuery(
+                // We do the JOIN here to restrict results to those present in the network map
+                "SELECT DISTINCT l FROM ${NodeInfoSchemaV1.PersistentNodeInfo::class.java.name} n JOIN n.legalIdentitiesAndCerts l WHERE l.name = :name",
+                NodeInfoSchemaV1.DBPartyAndCertificate::class.java)
+        query.setParameter("name", name.toString())
+        val candidates = query.resultList.map { it.toLegalIdentityAndCert() }
+        // The map is restricted to holding a single identity for any X.500 name, so firstOrNull() is correct here.
+        return candidates.firstOrNull()
     }
 
-    private fun queryByLegalName(name: CordaX500Name): List<NodeInfo> {
-        createSession {
-            val query = it.createQuery(
-                    "SELECT n FROM ${NodeInfoSchemaV1.PersistentNodeInfo::class.java.name} n JOIN n.legalIdentitiesAndCerts l WHERE l.name = :name",
-                    NodeInfoSchemaV1.PersistentNodeInfo::class.java)
-            query.setParameter("name", name.toString())
-            val result = query.resultList
-            return result.map { it.toNodeInfo() }
-        }
+    private fun queryByLegalName(session: Session, name: CordaX500Name): List<NodeInfo> {
+        val query = session.createQuery(
+                "SELECT n FROM ${NodeInfoSchemaV1.PersistentNodeInfo::class.java.name} n JOIN n.legalIdentitiesAndCerts l WHERE l.name = :name",
+                NodeInfoSchemaV1.PersistentNodeInfo::class.java)
+        query.setParameter("name", name.toString())
+        val result = query.resultList
+        return result.map { it.toNodeInfo() }
     }
 
-    private fun queryByAddress(hostAndPort: NetworkHostAndPort): NodeInfo? {
-        createSession {
-            val query = it.createQuery(
-                    "SELECT n FROM ${NodeInfoSchemaV1.PersistentNodeInfo::class.java.name} n JOIN n.addresses a WHERE a.pk.host = :host AND a.pk.port = :port",
-                    NodeInfoSchemaV1.PersistentNodeInfo::class.java)
-            query.setParameter("host", hostAndPort.host)
-            query.setParameter("port", hostAndPort.port)
-            val result = query.resultList
-            return if (result.isEmpty()) null
-            else result.map { it.toNodeInfo() }.singleOrNull() ?: throw IllegalStateException("More than one node with the same host and port")
-        }
+    private fun queryByAddress(session: Session, hostAndPort: NetworkHostAndPort): NodeInfo? {
+        val query = session.createQuery(
+                "SELECT n FROM ${NodeInfoSchemaV1.PersistentNodeInfo::class.java.name} n JOIN n.addresses a WHERE a.pk.host = :host AND a.pk.port = :port",
+                NodeInfoSchemaV1.PersistentNodeInfo::class.java)
+        query.setParameter("host", hostAndPort.host)
+        query.setParameter("port", hostAndPort.port)
+        val result = query.resultList
+        return if (result.isEmpty()) null
+        else result.map { it.toNodeInfo() }.singleOrNull() ?: throw IllegalStateException("More than one node with the same host and port")
     }
 
     /** Object Relational Mapping support. */
@@ -388,10 +368,8 @@ open class PersistentNetworkMapCache(private val serviceHub: ServiceHubInternal)
 
     override fun clearNetworkMapCache() {
         serviceHub.database.transaction {
-            createSession {
-                val result = getAllInfos(it)
-                for (nodeInfo in result) it.remove(nodeInfo)
-            }
+            val result = getAllInfos(session)
+            for (nodeInfo in result) session.remove(nodeInfo)
         }
     }
 }