+++ tools/demobench/src/main/kotlin/net/corda/demobench/model/HasCordapps.kt
@@ -2,6 +2,6 @@ package net.corda.demobench.model
 
 import java.nio.file.Path
 
-interface HasPlugins {
-    val pluginDir: Path
+interface HasCordapps {
+    val cordappsDir: Path
 }